Convert RageshakeDataStore to an interface for testing purpose.
