+++ features/rageshake/src/main/kotlin/io/element/android/features/rageshake/rageshake/RageshakeDataStore.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2022 New Vector Ltd
+ * Copyright (c) 2023 New Vector Ltd
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -16,54 +16,16 @@
 
 package io.element.android.features.rageshake.rageshake
 
-import android.content.Context
-import androidx.datastore.core.DataStore
-import androidx.datastore.preferences.core.Preferences
-import androidx.datastore.preferences.core.booleanPreferencesKey
-import androidx.datastore.preferences.core.edit
-import androidx.datastore.preferences.core.floatPreferencesKey
-import androidx.datastore.preferences.preferencesDataStore
-import io.element.android.libraries.core.bool.orTrue
-import io.element.android.libraries.di.ApplicationContext
 import kotlinx.coroutines.flow.Flow
-import kotlinx.coroutines.flow.map
-import javax.inject.Inject
 
-private val Context.dataStore: DataStore<Preferences> by preferencesDataStore(name = "elementx_rageshake")
+interface RageshakeDataStore {
+    fun isEnabled(): Flow<Boolean>
 
-private val enabledKey = booleanPreferencesKey("enabled")
-private val sensitivityKey = floatPreferencesKey("sensitivity")
+    suspend fun setIsEnabled(isEnabled: Boolean)
 
-class RageshakeDataStore @Inject constructor(
-    @ApplicationContext context: Context
-) {
-    private val store = context.dataStore
+    fun sensitivity(): Flow<Float>
 
-    fun isEnabled(): Flow<Boolean> {
-        return store.data.map { prefs ->
-            prefs[enabledKey].orTrue()
-        }
-    }
+    suspend fun setSensitivity(sensitivity: Float)
 
-    suspend fun setIsEnabled(isEnabled: Boolean) {
-        store.edit { prefs ->
-            prefs[enabledKey] = isEnabled
-        }
-    }
-
-    fun sensitivity(): Flow<Float> {
-        return store.data.map { prefs ->
-            prefs[sensitivityKey] ?: 0.5f
-        }
-    }
-
-    suspend fun setSensitivity(sensitivity: Float) {
-        store.edit { prefs ->
-            prefs[sensitivityKey] = sensitivity
-        }
-    }
-
-    suspend fun reset() {
-        store.edit { it.clear() }
-    }
+    suspend fun reset()
 }
